CheatSheetAI V3/cheatsheet_ai/requirements.txt
python-dotenv==0.19.2
pydantic==1.9.0
langchain==0.0.184
langchain-openai==0.0.2
langchain-anthropic==0.0.2
openai==0.27.0
anthropic==0.2.6
browser-use==0.1.0
asyncio==3.4.3
